Move elementDefinitionIconConverter functionality to image node selector
