+++ EngineeringModel/Selectors/ElementDefinitionTreeListNodeImageSelector.cs
@@ -25,11 +25,19 @@
 
 namespace CDP4EngineeringModel.Selectors
 {
+    using System;
     using System.Globalization;
+    using System.Linq;
+    using System.Windows.Data;
     using System.Windows.Media;
+    using System.Windows.Media.Imaging;
 
     using CDP4Common.CommonData;
+    using CDP4Common.EngineeringModelData;
+    using CDP4Common.Helpers;
+    using CDP4Common.SiteDirectoryData;
 
+    using CDP4Composition;
     using CDP4Composition.Mvvm;
     using CDP4Composition.Services;
 
@@ -69,11 +77,70 @@ namespace CDP4EngineeringModel.Selectors
             {
                 classKindOverride = ClassKind.ParameterTypeComponent;
             }
-
-            var elementDefinitionBrowserIconConverter = new ElementDefinitionBrowserIconConverter();
-            var image = elementDefinitionBrowserIconConverter.Convert(new object[] { thingStatus }, null, classKindOverride, CultureInfo.InvariantCulture);
+            
+            var image = this.Convert(new object[] { thingStatus }, null, classKindOverride, CultureInfo.InvariantCulture);
 
             return image as ImageSource;
+        }
+
+        /// <summary>
+        /// Returns an GetImage (icon) based on the <see cref="Thing" /> that is provided
+        /// </summary>
+        /// <param name="value">An instance of <see cref="Thing" /> for which an Icon needs to be returned</param>
+        /// <param name="targetType">The parameter is not used.</param>
+        /// <param name="parameter">The <see cref="ClassKind" /> of the overlay to use</param>
+        /// <param name="culture">The parameter is not used.</param>
+        /// <returns>
+        /// A <see cref="Uri" /> to an GetImage
+        /// </returns>
+        public object Convert(object[] value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var genericConverter = new ThingToIconUriConverter();
+            var thingStatus = value.SingleOrDefault() as ThingStatus;
+
+            var parameterBase = thingStatus?.Thing as ParameterBase;
+
+            ClassKind valuesetRowType;
+
+            if (parameterBase == null || parameter == null || !Enum.TryParse(parameter.ToString(), out valuesetRowType))
+            {
+                return genericConverter.Convert(value, targetType, parameter, culture);
+            }
+
+            var isCompound = parameterBase.ParameterType is CompoundParameterType;
+
+            // Value set row
+            // row representing an option
+            if (valuesetRowType == ClassKind.Option)
+            {
+                var optionUri = new Uri(IconUtilities.ImageUri(valuesetRowType).ToString());
+
+                if (parameterBase.StateDependence != null || isCompound)
+                {
+                    return new BitmapImage(optionUri);
+                }
+
+                var uri = new Uri(IconUtilities.ImageUri(parameterBase.ClassKind).ToString());
+                return IconUtilities.WithOverlay(uri, optionUri);
+            }
+
+            // row representing a component
+            if (valuesetRowType == ClassKind.ParameterTypeComponent)
+            {
+                var componentUri = new Uri(IconUtilities.ImageUri(valuesetRowType).ToString());
+                return new BitmapImage(componentUri);
+            }
+
+            // Row representing state
+            var stateUri = new Uri(IconUtilities.ImageUri(ClassKind.ActualFiniteState).ToString());
+
+            if (isCompound)
+            {
+                return new BitmapImage(stateUri);
+            }
+
+            var baseUri = new Uri(IconUtilities.ImageUri(parameterBase.ClassKind).ToString());
+            return IconUtilities.WithOverlay(baseUri, stateUri);
         }
     }
 }
